added /join POST handler, static /pages/ directory
The post handler sends the user POSTed data to the logs as well as
storing it in the datastore and sending an email to a admin email
address which is configurable in config.py.

The static /pages/ dir was setup to serve any static files we might want
to post up on our domain. They are served as the mime type indicated by
the file extension.

 On branch version_3
	modified:   google-app-engine/app.yaml
	modified:   google-app-engine/config.py
	modified:   google-app-engine/request.py
	new file:   google-app-engine/static/pages/extraordinary_people.html